Make some text a little bolder in Chrome
Fixed 4 instances of too light font by bumping up the font-weight by 1
(this seems to operate as a step-function and not continuous):
* The date next to posts
* The tags next to posts
* Tags in the tags page
* Pagination forward/back links

Note that this only affects Chrome; Safari and Firefox had zero visual
changes. Blockquotes had no changes when bumping up the font-weight,
and the site header and site navigation links were a little too bold in
Chrome, which is fine since the fonts are a bit bigger to compensate.

This all happens because we use the Source Sans Pro font. This issue
didn't manifest as heavily in the sample Spoqa Han Sans font.